# specs/007-redis-ratelimit-openapi/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: Nest Platform API
  description: |
    API for the Nest multi-agent real estate platform.

    ## Authentication
    - **Admin/Agent routes**: Require Supabase session cookie
    - **Public routes**: No authentication required
    - **Webhook routes**: Require webhook signature validation

    ## Rate Limiting
    - Login: 5 attempts per email per 15 minutes
    - Contact form: 5 requests per IP per hour
  version: 1.0.0
  contact:
    name: Nest Associates
servers:
  - url: https://dashboard.nestassociates.com
    description: Production
  - url: http://localhost:3000
    description: Development
...
paths:
...
  /api/agent/profile:
    get:
      tags: [Agent]
      summary: Get current agent profile
      security:
        - cookieAuth: []
      responses:
        '200':
          description: Agent profile
    patch:
      tags: [Agent]
      summary: Update agent profile
      security:
        - cookieAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateProfileRequest'
      responses:
        '200':
          description: Profile updated
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: sb-access-token
      description: Supabase session cookie
  schemas:
...
    UpdateProfileRequest:
      type: object
      properties:
        phone:
          type: string
        bio:
          type: string
        avatar_url:
          type: string
          format: uri

    # Content
...